Move App.tsx to components directory
