+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import App from '@/App.tsx';
+import App from '@/components/App.tsx';
 
 import './main.css';
